end to end tests for the game page
