+++ client/src/modules/front_end_components/ModalManager.js
