+++ client/src/scripts/howToUse.js
@@ -1,4 +1,4 @@
-import { injectNavbar } from '../modules/Navbar.js';
+import { injectNavbar } from '../modules/front_end_components/Navbar.js';
 
 const howToUse = () => {
     injectNavbar();
