+++ client/src/views/game.html
@@ -19,46 +19,8 @@
     <link rel="preload" href="/webfonts/SignikaNegative-Light.woff2" as="font" type="font/woff2" crossorigin>
 </head>
 <body>
-<div id="role-info-prompt"></div>
-<div id="transfer-mod-prompt"></div>
-<div class="spinner-background"></div>
-<div class="spinner-container">
-    <div class="lds-spinner">
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-    </div>
-    <p>Connecting to game...</p>
-</div>
-<div id="mobile-menu-background-overlay"></div>
-<div id="navbar"></div>
-<div id="game-content" class="container">
-    <div class="placeholder-row">
-        <div class="animated-placeholder animated-placeholder-short"></div>
-        <div class="animated-placeholder animated-placeholder-short animated-placeholder-invisible"></div>
-    </div>
-    <div class="placeholder-row">
-        <div class="animated-placeholder animated-placeholder-short"></div>
-        <div class="animated-placeholder animated-placeholder-short animated-placeholder-invisible"></div>
-    </div>
-    <div class="animated-placeholder animated-placeholder-long"></div>
-    <div class="animated-placeholder animated-placeholder-long"></div>
-    <div class="placeholder-row">
-        <div class="animated-placeholder animated-placeholder-short"></div>
-        <div class="animated-placeholder animated-placeholder-short animated-placeholder-invisible"></div>
-    </div>
-    <div class="animated-placeholder animated-placeholder-long"></div>
-</div>
-<script src="/dist/game-bundle.js"></script>
+    <script src="/dist/game-bundle.js"></script>
+    <script src="/socket.io/socket.io.js"></script>
 </body>
 </html>
