+++ package.json
@@ -25,7 +25,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@babel/core": "^7.16.7",
+    "@babel/core": "^7.18.13",
     "@babel/eslint-parser": "^7.16.5",
     "@babel/plugin-transform-object-assign": "^7.16.5",
     "@babel/preset-env": "^7.16.5",
@@ -33,6 +33,7 @@
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.2.3",
     "body-parser": "^1.19.1",
+    "core-js": "^3.25.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "express-force-https": "^1.0.0",
@@ -41,6 +42,7 @@
     "karma-jasmine": "^4.0.1",
     "open": "^7.0.3",
     "rate-limiter-flexible": "^2.3.6",
+    "regenerator-runtime": "^0.13.9",
     "socket.io": "^4.4.0",
     "socket.io-client": "^4.4.0",
     "webpack": "^5.65.0",
